other files/requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
faker
transformers
openpyxl
sqlalchemy 
pymysql
pyodbc
langchain
langchain_huggingface
spacy
rapidfuzz
google-cloud-aiplatform
gensim
nltk
genai
pyPDF2
keras_tuner